update mnist example

+++ agnez/grid.py
@@ -27,8 +27,8 @@ def img_grid(X, rows_cols=None, rescale=True):
     N, channels, height, width = X.shape
 
     if rows_cols is None:
-        cols = np.ceil(np.sqrt(X.shape[0]))
-        rows = np.ceil(X.shape[0] / cols)
+        cols = np.ceil(np.sqrt(X.shape[0])).astype('int')
+        rows = np.ceil(X.shape[0] / cols).astype('int')
     else:
         rows, cols = rows_cols
